# docs/apis/swagger.yaml
definitions:
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  models.Document:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
  title: Swagger REST API Documentation
  version: "1.0"
paths:
  /documents:
    get:
      description: Retrieve all documents
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Document'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Retrieve all documents
  /documents/{id}:
    delete:
      description: Delete a given document id
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Delete a given document id
    get:
      description: Retrieve  a given document from the path param id
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Document'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Retrieve a given document
    put:
      consumes:
      - application/json
      description: Create or update a document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      - description: The document struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Document'
      produces:
      - application/json
      responses:
        "200":
          description: update
          schema:
            $ref: '#/definitions/models.Document'
        "201":
          description: creation
          schema:
            $ref: '#/definitions/models.Document'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Create or update a document
  /emails:
    post:
      description: Post messages to kafka
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Post messages to kafka
swagger: "2.0"